Account backend (#42)
* Base entity

* Base repository

* Fixes

* Make repositories not static

* Test errors

* Repository fixes

* Slim down repositories

* Make authorisation blueprint more readable

* User dto

* Make user blueprint more readable

* Make user icon blueprint more readable

* Make transfer blueprint more readable

* Add dto for history of transactions

* Login response class

* User response class

* History response class

* analysis response class

* Update cyclic_payment_blueprint.py

* Make cyclic payment blueprint more readable

* Update user_icon_blueprint.py

* Fix login tests

* Fix register tests

* Fix get user tests

* Fix update user tests

* Fix update password tests

* Fix icon tests

* Fix create transfer tests

* Add transfer to self test

* Fix create loan tests

* Fix get transfer tests

* Fix monthly analysis tests

* Fix yearly analysis tests

* Fix create cyclic payment tests

* Fix delete cyclic payment tests

* Fix get cyclic payment tests

* Fix get cyclic payment list tests

* Fix put cyclic payment tests

* Small fixes

* Basic account implementation

* bank account

* fixes

* fixes

* fixes

* fixes

* Authorisation tests

* Authorisation blueprint

* Account tests

* Account blueprint

* User tests

* Transfers tests

* Cyclic payment tests

* Account backend tests (#43)

* Authentication tests

* Account tests

* User tests

* Transfer tests

* Cyclic payments tests

* Transfer tests and other chores

* Cyclic payments

* Add active account to user

* Fix transfers

* Fix loan

* Fix user

* Fix user again

* Add user data svg

* Fix accounts

* Add account data svg

* Fix transfers

* Partial cyclic payment fixes

* Fix cyclic payments

* Added update and delete account

* Added soft delete

* Remove active account when deleting account

* Delete user

* Logout after user deletion or password change

+++ backend/routes/cyclic_payment/delete_cyclic_payment.py
@@ -1,28 +1,39 @@
-from flask import Response
-from flask_login import login_required, current_user
+from http import HTTPStatus
 
-from cyclic_payments import CyclicPaymentRepository
+from flask import Response
+from flask_login import login_required
+
+from accounts import Account, AccountRepository
+from cyclic_payments import CyclicPaymentRepository, CyclicPayment
 from helpers import subtract
 from routes.cyclic_payment.helpers import validate_object_id
 from routes.helpers import create_simple_response
+from routes.transfer.helpers import prevent_unauthorised_account_access
 from users import UserRepository
 
 user_repository = UserRepository()
+account_repository = AccountRepository()
 cyclic_payment_repository = CyclicPaymentRepository()
 
 @login_required
-def delete_cyclic_payment(id) -> tuple[Response, int]:
+def delete_cyclic_payment(id) -> Response:
     error = validate_object_id(id)
     if error:
-        return create_simple_response(error, 400)
+        return create_simple_response(error, HTTPStatus.BAD_REQUEST)
 
-    cyclic_payment = cyclic_payment_repository.find_by_id(str(id))
+    cyclic_payment: CyclicPayment = cyclic_payment_repository.find_by_id(str(id))
     if not cyclic_payment:
-        return create_simple_response("cyclicPaymentNotExist", 404)
+        return create_simple_response("cyclicPaymentNotExist", HTTPStatus.NOT_FOUND)
 
-    curr_user = user_repository.find_by_id(current_user._id)
-    user_repository.update(curr_user.id, {'blockades': subtract(float(curr_user.blockades), float(cyclic_payment.amount))})
+    account: Account = account_repository.find_by_id(str(cyclic_payment.issuer_id))
+    if not account:
+        return create_simple_response("accountNotExist", HTTPStatus.NOT_FOUND)
+
+    if prevent_unauthorised_account_access(account):
+        return create_simple_response("unauthorisedAccountAccess", HTTPStatus.UNAUTHORIZED)
+
+    account_repository.update(str(account.id), {'blockades': subtract(float(account.blockades), float(cyclic_payment.amount))})
 
     cyclic_payment_repository.delete(str(id))
 
-    return create_simple_response("cyclicPaymentDeletedSuccessful", 200)
+    return create_simple_response("cyclicPaymentDeletedSuccessful", HTTPStatus.OK)
